Rename DefaultPropertyFactory.CodeGenerator -> DefaultProperty
DefaultPropertyFactory is now DefaultProperty.Factory
