+++ src/test/java/org/inferred/freebuilder/processor/RequiredPropertiesSourceTest.java
@@ -1697,10 +1697,10 @@ public class RequiredPropertiesSourceTest {
     return metadata.toBuilder()
         .clearProperties()
         .addProperties(name.toBuilder()
-            .setCodeGenerator(new DefaultPropertyFactory.CodeGenerator(metadata, name, false))
+            .setCodeGenerator(new DefaultProperty(metadata, name, false))
             .build())
         .addProperties(age.toBuilder()
-            .setCodeGenerator(new DefaultPropertyFactory.CodeGenerator(metadata, age, false))
+            .setCodeGenerator(new DefaultProperty(metadata, age, false))
             .build())
         .build();
   }
